Added webserver requirement due to CORS policy
CORS policy requires reading file requests to come from http:// and does not allow local disk directory. Because of that I decided to implement a small ajax request method to read and write to the JSON file. I recommend you use xampp in order to be able to launch the web app.